+++ packages/rum/test/userActionCollection.spec.ts
@@ -1,49 +1,24 @@
-import { DOM_EVENT, ErrorMessage, Observable, RequestCompleteEvent } from '@keitoaino/datadog-browser-core'
+import { DOM_EVENT, ErrorMessage } from '@keitoaino/datadog-browser-core'
 import { LifeCycle, LifeCycleEventType } from '../src/lifeCycle'
+import { PAGE_ACTIVITY_MAX_DURATION, PAGE_ACTIVITY_VALIDATION_DELAY } from '../src/trackPageActivities'
 import {
   $$tests,
   AutoUserAction,
   getUserActionReference,
-  PageActivityEvent,
-  startUserActionCollection,
-  USER_ACTION_END_DELAY,
-  USER_ACTION_MAX_DURATION,
-  USER_ACTION_VALIDATION_DELAY,
   UserAction,
   UserActionType,
 } from '../src/userActionCollection'
-const { waitUserActionCompletion, trackPageActivities, resetUserAction, newUserAction } = $$tests
+import { View } from '../src/viewCollection'
+import { setup, TestSetupBuilder } from './specHelper'
+
+const { resetUserAction, newUserAction } = $$tests
 
 // Used to wait some time after the creation of a user action
-const BEFORE_USER_ACTION_VALIDATION_DELAY = USER_ACTION_VALIDATION_DELAY * 0.8
-// Used to wait some time before the (potential) end of a user action
-const BEFORE_USER_ACTION_END_DELAY = USER_ACTION_END_DELAY * 0.8
+const BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY = PAGE_ACTIVITY_VALIDATION_DELAY * 0.8
 // Used to wait some time but it doesn't matter how much.
 const SOME_ARBITRARY_DELAY = 50
 // A long delay used to wait after any user action is finished.
-const EXPIRE_DELAY = USER_ACTION_MAX_DURATION * 10
-
-function mockClock() {
-  beforeEach(() => {
-    jasmine.clock().install()
-    jasmine.clock().mockDate()
-    spyOn(performance, 'now').and.callFake(() => Date.now())
-  })
-
-  afterEach(() => {
-    jasmine.clock().uninstall()
-  })
-
-  return {
-    tick(ms: number) {
-      jasmine.clock().tick(ms)
-    },
-    expire() {
-      // Make sure no user action is still pending
-      jasmine.clock().tick(EXPIRE_DELAY)
-    },
-  }
-}
+const EXPIRE_DELAY = PAGE_ACTIVITY_MAX_DURATION * 10
 
 function eventsCollector<T>() {
   const events: T[] = []
@@ -60,30 +35,12 @@ function eventsCollector<T>() {
 
 describe('startUserActionCollection', () => {
   const { events, pushEvent } = eventsCollector()
-  const clock = mockClock()
   let button: HTMLButtonElement
-  let userActionCollectionSubscription: { stop(): void }
-  let lifeCycle: LifeCycle
+  let setupBuilder: TestSetupBuilder
 
-  beforeEach(() => {
-    button = document.createElement('button')
-    button.type = 'button'
-    button.appendChild(document.createTextNode('Click me'))
-    document.body.appendChild(button)
-
-    lifeCycle = new LifeCycle()
-    lifeCycle.subscribe(LifeCycleEventType.USER_ACTION_COLLECTED, pushEvent)
-    userActionCollectionSubscription = startUserActionCollection(lifeCycle)
-  })
-
-  afterEach(() => {
-    button.parentNode!.removeChild(button)
-    userActionCollectionSubscription.stop()
-  })
-
-  it('starts a user action when clicking on an element', () => {
+  function mockValidatedClickUserAction(lifeCycle: LifeCycle, clock: jasmine.Clock) {
     button.addEventListener(DOM_EVENT.CLICK, () => {
-      clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
+      clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
       // Since we don't collect dom mutations for this test, manually dispatch one
       lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
     })
@@ -91,10 +48,50 @@ describe('startUserActionCollection', () => {
     clock.tick(SOME_ARBITRARY_DELAY)
     button.click()
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
+  }
+
+  beforeEach(() => {
+    button = document.createElement('button')
+    button.type = 'button'
+    button.appendChild(document.createTextNode('Click me'))
+    document.body.appendChild(button)
+
+    setupBuilder = setup()
+      .withFakeClock()
+      .withUserActionCollection()
+      .beforeBuild((lifeCycle) => lifeCycle.subscribe(LifeCycleEventType.USER_ACTION_COLLECTED, pushEvent))
+  })
+
+  afterEach(() => {
+    button.parentNode!.removeChild(button)
+    setupBuilder.cleanup()
+  })
+
+  it('cancels pending user action on view loading', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+
+    button.addEventListener(DOM_EVENT.CLICK, () => {
+      clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+      // Since we don't collect dom mutations for this test, manually dispatch one
+      lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    })
+
+    clock.tick(SOME_ARBITRARY_DELAY)
+    button.click()
+    const fakeView = {}
+    lifeCycle.notify(LifeCycleEventType.VIEW_COLLECTED, fakeView as View)
+    clock.tick(EXPIRE_DELAY)
+
+    expect(events).toEqual([])
+  })
+
+  it('starts a user action when clicking on an element', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+    mockValidatedClickUserAction(lifeCycle, clock)
     expect(events).toEqual([
       {
-        duration: BEFORE_USER_ACTION_VALIDATION_DELAY,
+        duration: BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY,
         id: jasmine.any(String),
         measures: {
           errorCount: 0,
@@ -109,23 +106,30 @@ describe('startUserActionCollection', () => {
   })
 
   it('cancels a user action when if nothing happens after a click', () => {
+    const { clock } = setupBuilder.build()
     clock.tick(SOME_ARBITRARY_DELAY)
     button.click()
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
     expect(events).toEqual([])
   })
 })
 
 describe('getUserActionReference', () => {
-  const clock = mockClock()
+  let setupBuilder: TestSetupBuilder
   const { events, pushEvent } = eventsCollector<UserAction>()
 
   beforeEach(() => {
+    setupBuilder = setup().withFakeClock()
+  })
+
+  afterEach(() => {
     resetUserAction()
+    setupBuilder.cleanup()
   })
 
   it('returns the current user action reference', () => {
+    const { clock } = setupBuilder.build()
     expect(getUserActionReference()).toBeUndefined()
     const lifeCycle = new LifeCycle()
     lifeCycle.subscribe(LifeCycleEventType.USER_ACTION_COLLECTED, pushEvent)
@@ -136,12 +140,12 @@ describe('getUserActionReference', () => {
 
     expect(userActionReference).toBeDefined()
 
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
     lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
 
     expect(getUserActionReference()).toBeDefined()
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
 
     expect(getUserActionReference()).toBeUndefined()
 
@@ -150,55 +154,64 @@ describe('getUserActionReference', () => {
   })
 
   it('do not return the user action reference for events occuring before the start of the user action', () => {
+    const { clock } = setupBuilder.build()
     const timeBeforeStartingUserAction = Date.now()
 
     clock.tick(SOME_ARBITRARY_DELAY)
     newUserAction(new LifeCycle(), UserActionType.CLICK, 'test')
 
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY * 0.5)
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY * 0.5)
     const timeAfterStartingUserAction = Date.now()
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY * 0.5)
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY * 0.5)
 
     expect(getUserActionReference()).toBeDefined()
     expect(getUserActionReference(timeAfterStartingUserAction)).toBeDefined()
     expect(getUserActionReference(timeBeforeStartingUserAction)).toBeUndefined()
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
   })
 })
 
 describe('newUserAction', () => {
-  const clock = mockClock()
+  let setupBuilder: TestSetupBuilder
   const { events, pushEvent } = eventsCollector<UserAction>()
 
+  beforeEach(() => {
+    setupBuilder = setup().withFakeClock()
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
+  })
+
   it('cancels any starting user action while another one is happening', () => {
-    const lifeCycle = new LifeCycle()
+    const { lifeCycle, clock } = setupBuilder.build()
     lifeCycle.subscribe(LifeCycleEventType.USER_ACTION_COLLECTED, pushEvent)
 
     newUserAction(lifeCycle, UserActionType.CLICK, 'test-1')
     newUserAction(lifeCycle, UserActionType.CLICK, 'test-2')
 
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
     lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
     expect(events.length).toBe(1)
     expect(events[0].name).toBe('test-1')
   })
 
   it('counts errors occuring during the user action', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
     const error = {}
-    const lifeCycle = new LifeCycle()
     lifeCycle.subscribe(LifeCycleEventType.USER_ACTION_COLLECTED, pushEvent)
 
     newUserAction(lifeCycle, UserActionType.CLICK, 'test-1')
 
     lifeCycle.notify(LifeCycleEventType.ERROR_COLLECTED, error as ErrorMessage)
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
     lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
     lifeCycle.notify(LifeCycleEventType.ERROR_COLLECTED, error as ErrorMessage)
 
-    clock.expire()
+    clock.tick(EXPIRE_DELAY)
     lifeCycle.notify(LifeCycleEventType.ERROR_COLLECTED, error as ErrorMessage)
 
     expect(events.length).toBe(1)
@@ -207,203 +220,6 @@ describe('newUserAction', () => {
       errorCount: 2,
       longTaskCount: 0,
       resourceCount: 0,
-    })
-  })
-})
-
-describe('trackPagePageActivities', () => {
-  const { events, pushEvent } = eventsCollector<PageActivityEvent>()
-  it('emits an activity event on dom mutation', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
-    expect(events).toEqual([{ isBusy: false }])
-  })
-
-  it('emits an activity event on resource collected', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-    const performanceEntry = {
-      entryType: 'resource',
-    }
-    lifeCycle.notify(LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED, performanceEntry as PerformanceEntry)
-    expect(events).toEqual([{ isBusy: false }])
-  })
-
-  it('does not emit an activity event when a navigation occurs', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-    const performanceEntry = {
-      entryType: 'navigation',
-    }
-    lifeCycle.notify(LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED, performanceEntry as PerformanceEntry)
-    expect(events).toEqual([])
-  })
-
-  it('stops emiting activities after calling stop()', () => {
-    const lifeCycle = new LifeCycle()
-    const { stop, observable } = trackPageActivities(lifeCycle)
-    observable.subscribe(pushEvent)
-
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
-    expect(events).toEqual([{ isBusy: false }])
-
-    stop()
-
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
-
-    expect(events).toEqual([{ isBusy: false }])
-  })
-
-  describe('requests', () => {
-    function makeFakeRequestCompleteEvent(requestId: number): RequestCompleteEvent {
-      return { requestId } as any
-    }
-    it('emits an activity event when a request starts', () => {
-      const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-      lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
-        requestId: 10,
-      })
-      expect(events).toEqual([{ isBusy: true }])
-    })
-
-    it('emits an activity event when a request completes', () => {
-      const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-      lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
-        requestId: 10,
-      })
-      lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(10))
-      expect(events).toEqual([{ isBusy: true }, { isBusy: false }])
-    })
-
-    it('ignores requests that has started before', () => {
-      const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-      lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(10))
-      expect(events).toEqual([])
-    })
-
-    it('keeps emiting busy events while all requests are not completed', () => {
-      const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-      lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
-        requestId: 10,
-      })
-      lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
-        requestId: 11,
-      })
-      lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(9))
-      lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(11))
-      lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(10))
-      expect(events).toEqual([{ isBusy: true }, { isBusy: true }, { isBusy: true }, { isBusy: false }])
-    })
-  })
-})
-
-describe('waitUserActionCompletion', () => {
-  const clock = mockClock()
-
-  it('should not collect an event that is not followed by page activity', (done) => {
-    waitUserActionCompletion(new Observable(), (endTime) => {
-      expect(endTime).toBeUndefined()
-      done()
-    })
-
-    clock.expire()
-  })
-
-  it('should collect an event that is followed by page activity', (done) => {
-    const activityObservable = new Observable<PageActivityEvent>()
-
-    const startTime = performance.now()
-    waitUserActionCompletion(activityObservable, (endTime) => {
-      expect(endTime).toEqual(startTime + BEFORE_USER_ACTION_VALIDATION_DELAY)
-      done()
-    })
-
-    clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
-    activityObservable.notify({ isBusy: false })
-
-    clock.expire()
-  })
-
-  describe('extend with activities', () => {
-    it('is extended while there is page activities', (done) => {
-      const activityObservable = new Observable<PageActivityEvent>()
-      const startTime = performance.now()
-
-      // Extend the user action but stops before USER_ACTION_MAX_DURATION
-      const extendCount = Math.floor(USER_ACTION_MAX_DURATION / BEFORE_USER_ACTION_END_DELAY - 1)
-
-      waitUserActionCompletion(activityObservable, (endTime) => {
-        expect(endTime).toBe(startTime + (extendCount + 1) * BEFORE_USER_ACTION_END_DELAY)
-        done()
-      })
-
-      for (let i = 0; i < extendCount; i += 1) {
-        clock.tick(BEFORE_USER_ACTION_END_DELAY)
-        activityObservable.notify({ isBusy: false })
-      }
-
-      clock.expire()
-    })
-
-    it('expires after a limit', (done) => {
-      const activityObservable = new Observable<PageActivityEvent>()
-      let stop = false
-      const startTime = performance.now()
-
-      // Extend the user action until it's more than USER_ACTION_MAX_DURATION
-      const extendCount = Math.ceil(USER_ACTION_MAX_DURATION / BEFORE_USER_ACTION_END_DELAY + 1)
-
-      waitUserActionCompletion(activityObservable, (endTime) => {
-        expect(endTime).toBe(startTime + USER_ACTION_MAX_DURATION)
-        stop = true
-        done()
-      })
-
-      for (let i = 0; i < extendCount && !stop; i += 1) {
-        clock.tick(BEFORE_USER_ACTION_END_DELAY)
-        activityObservable.notify({ isBusy: false })
-      }
-
-      clock.expire()
-    })
-  })
-
-  describe('busy activities', () => {
-    it('is extended while the page is busy', (done) => {
-      const activityObservable = new Observable<PageActivityEvent>()
-      const startTime = performance.now()
-      waitUserActionCompletion(activityObservable, (endTime) => {
-        expect(endTime).toBe(startTime + BEFORE_USER_ACTION_VALIDATION_DELAY + USER_ACTION_END_DELAY * 2)
-        done()
-      })
-
-      clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
-      activityObservable.notify({ isBusy: true })
-
-      clock.tick(USER_ACTION_END_DELAY * 2)
-      activityObservable.notify({ isBusy: false })
-
-      clock.expire()
-    })
-
-    it('expires is the page is busy for too long', (done) => {
-      const activityObservable = new Observable<PageActivityEvent>()
-      const startTime = performance.now()
-      waitUserActionCompletion(activityObservable, (endTime) => {
-        expect(endTime).toBe(startTime + USER_ACTION_MAX_DURATION)
-        done()
-      })
-
-      clock.tick(BEFORE_USER_ACTION_VALIDATION_DELAY)
-      activityObservable.notify({ isBusy: true })
-
-      clock.expire()
     })
   })
 })