+++ packages/rum/test/viewCollection.spec.ts
@@ -1,35 +1,100 @@
 import { getHash, getPathName, getSearch } from '@keitoaino/datadog-browser-core'
 
-import { LifeCycle, LifeCycleEventType } from '../src/lifeCycle'
+import { LifeCycleEventType } from '../src/lifeCycle'
 import { PerformanceLongTaskTiming, PerformancePaintTiming } from '../src/rum'
-import { RumSession } from '../src/rumSession'
-import { UserAction, UserActionType } from '../src/userActionCollection'
-import { startViewCollection, THROTTLE_VIEW_UPDATE_PERIOD, View, viewContext } from '../src/viewCollection'
 
-function setup(lifeCycle: LifeCycle = new LifeCycle()) {
+import {
+  PAGE_ACTIVITY_END_DELAY,
+  PAGE_ACTIVITY_MAX_DURATION,
+  PAGE_ACTIVITY_VALIDATION_DELAY,
+} from '../src/trackPageActivities'
+import { UserAction, UserActionType } from '../src/userActionCollection'
+import { THROTTLE_VIEW_UPDATE_PERIOD, View, viewContext, ViewLoadingType } from '../src/viewCollection'
+import { setup, TestSetupBuilder } from './specHelper'
+
+const AFTER_PAGE_ACTIVITY_MAX_DURATION = PAGE_ACTIVITY_MAX_DURATION * 1.1
+const BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY = PAGE_ACTIVITY_VALIDATION_DELAY * 0.8
+const AFTER_PAGE_ACTIVITY_END_DELAY = PAGE_ACTIVITY_END_DELAY * 1.1
+
+const FAKE_LONG_TASK = {
+  entryType: 'longtask',
+  startTime: 456,
+}
+const FAKE_USER_ACTION = {
+  context: {
+    bar: 123,
+  },
+  name: 'foo',
+  type: UserActionType.CUSTOM,
+}
+const FAKE_PAINT_ENTRY = {
+  entryType: 'paint',
+  name: 'first-contentful-paint',
+  startTime: 123,
+}
+const FAKE_NAVIGATION_ENTRY = {
+  domComplete: 456,
+  domContentLoadedEventEnd: 345,
+  domInteractive: 234,
+  entryType: 'navigation',
+  loadEventEnd: 567,
+}
+
+const FAKE_NAVIGATION_ENTRY_WITH_LOADEVENT_BEFORE_ACTIVITY_TIMING = {
+  domComplete: 2,
+  domContentLoadedEventEnd: 1,
+  domInteractive: 1,
+  entryType: 'navigation',
+  loadEventEnd: BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY * 0.8,
+}
+
+const FAKE_NAVIGATION_ENTRY_WITH_LOADEVENT_AFTER_ACTIVITY_TIMING = {
+  domComplete: 2,
+  domContentLoadedEventEnd: 1,
+  domInteractive: 1,
+  entryType: 'navigation',
+  loadEventEnd: BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY * 1.2,
+}
+
+function mockHistory(location: Partial<Location>) {
   spyOn(history, 'pushState').and.callFake((_: any, __: string, pathname: string) => {
     const url = `http://localhost${pathname}`
-    fakeLocation.pathname = getPathName(url)
-    fakeLocation.search = getSearch(url)
-    fakeLocation.hash = getHash(url)
+    location.pathname = getPathName(url)
+    location.search = getSearch(url)
+    location.hash = getHash(url)
   })
-  const fakeLocation: Partial<Location> = { pathname: '/foo' }
-  const fakeSession = {
-    getId() {
-      return '42'
-    },
+}
+
+function spyOnViews() {
+  const addRumEvent = jasmine.createSpy()
+
+  function getViewEvent(index: number) {
+    return addRumEvent.calls.argsFor(index)[0] as View
   }
-  startViewCollection(fakeLocation as Location, lifeCycle, fakeSession as RumSession)
+
+  function getRumEventCount() {
+    return addRumEvent.calls.count()
+  }
+
+  return { addRumEvent, getViewEvent, getRumEventCount }
 }
 
 describe('rum track url change', () => {
+  let setupBuilder: TestSetupBuilder
   let initialView: string
   let initialLocation: Location
 
   beforeEach(() => {
-    setup()
+    const fakeLocation: Partial<Location> = { pathname: '/foo' }
+    mockHistory(fakeLocation)
+    setupBuilder = setup().withViewCollection(fakeLocation)
+    setupBuilder.build()
     initialView = viewContext.id
     initialLocation = viewContext.location
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
   })
 
   it('should update view id on path change', () => {
@@ -54,35 +119,28 @@ describe('rum track url change', () => {
   })
 })
 
-function spyOnViews() {
-  let addRumEvent: jasmine.Spy
-  let lifeCycle: LifeCycle
-
-  function getViewEvent(index: number) {
-    return addRumEvent.calls.argsFor(index)[0] as View
-  }
-
-  function getRumEventCount() {
-    return addRumEvent.calls.count()
-  }
-
-  addRumEvent = jasmine.createSpy()
-  lifeCycle = new LifeCycle()
-  lifeCycle.subscribe(LifeCycleEventType.VIEW_COLLECTED, addRumEvent)
-  setup(lifeCycle)
-
-  return { lifeCycle, getViewEvent, getRumEventCount }
-}
-
 describe('rum track renew session', () => {
-  let lifeCycle: LifeCycle
+  let setupBuilder: TestSetupBuilder
+  let addRumEvent: jasmine.Spy
   let getRumEventCount: () => number
   let getViewEvent: (index: number) => View
+
   beforeEach(() => {
-    ;({ lifeCycle, getRumEventCount, getViewEvent } = spyOnViews())
+    ;({ addRumEvent, getViewEvent, getRumEventCount } = spyOnViews())
+
+    const fakeLocation: Partial<Location> = { pathname: '/foo' }
+    mockHistory(fakeLocation)
+    setupBuilder = setup()
+      .withViewCollection(fakeLocation)
+      .beforeBuild((lifeCycle) => lifeCycle.subscribe(LifeCycleEventType.VIEW_COLLECTED, addRumEvent))
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
   })
 
   it('should update page view id on renew session', () => {
+    const { lifeCycle } = setupBuilder.build()
     const initialView = viewContext.id
     lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
 
@@ -90,6 +148,7 @@ describe('rum track renew session', () => {
   })
 
   it('should send a final view event when the session is renewed', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
 
     lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
@@ -99,38 +158,160 @@ describe('rum track renew session', () => {
   })
 })
 
+describe('rum track load duration', () => {
+  let setupBuilder: TestSetupBuilder
+  let addRumEvent: jasmine.Spy
+  let getViewEvent: (index: number) => View
+
+  beforeEach(() => {
+    ;({ addRumEvent, getViewEvent } = spyOnViews())
+
+    const fakeLocation: Partial<Location> = { pathname: '/foo' }
+    mockHistory(fakeLocation)
+    setupBuilder = setup()
+      .withFakeClock()
+      .withViewCollection(fakeLocation)
+      .beforeBuild((lifeCycle) => lifeCycle.subscribe(LifeCycleEventType.VIEW_COLLECTED, addRumEvent))
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
+  })
+
+  it('should collect intital view type as "initial_load"', () => {
+    setupBuilder.build()
+    expect(getViewEvent(0).loadingType).toEqual(ViewLoadingType.INITIAL_LOAD)
+  })
+
+  it('should collect view type as "route_change" after a route change', () => {
+    setupBuilder.build()
+    history.pushState({}, '', '/bar')
+    expect(getViewEvent(1).location.pathname).toEqual('/foo')
+    expect(getViewEvent(1).loadingType).toEqual(ViewLoadingType.INITIAL_LOAD)
+
+    expect(getViewEvent(2).location.pathname).toEqual('/bar')
+    expect(getViewEvent(2).loadingType).toEqual(ViewLoadingType.ROUTE_CHANGE)
+  })
+})
+
+describe('rum track loading time', () => {
+  let setupBuilder: TestSetupBuilder
+  let addRumEvent: jasmine.Spy
+  let getViewEvent: (index: number) => View
+  let getRumEventCount: () => number
+
+  beforeEach(() => {
+    ;({ addRumEvent, getRumEventCount, getViewEvent } = spyOnViews())
+
+    const fakeLocation: Partial<Location> = { pathname: '/foo' }
+    mockHistory(fakeLocation)
+    setupBuilder = setup()
+      .withFakeClock()
+      .withViewCollection(fakeLocation)
+      .beforeBuild((lifeCycle) => lifeCycle.subscribe(LifeCycleEventType.VIEW_COLLECTED, addRumEvent))
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
+  })
+
+  it('should have an undefined loading time if there is no activity on a route change', () => {
+    const { clock } = setupBuilder.build()
+
+    history.pushState({}, '', '/bar')
+    clock.tick(AFTER_PAGE_ACTIVITY_MAX_DURATION)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(getRumEventCount()).toEqual(3)
+    expect(getViewEvent(2).loadingTime).toBeUndefined()
+  })
+
+  it('should have a loading time equal to the activity time if there is a unique activity on a route change', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+
+    history.pushState({}, '', '/bar')
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    clock.tick(AFTER_PAGE_ACTIVITY_END_DELAY)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(getViewEvent(3).loadingTime).toEqual(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+  })
+
+  it('should use loadEventEnd for initial view when having no activity', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+    expect(getRumEventCount()).toEqual(1)
+
+    lifeCycle.notify(
+      LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED,
+      FAKE_NAVIGATION_ENTRY as PerformanceNavigationTiming
+    )
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(getRumEventCount()).toEqual(2)
+    expect(getViewEvent(1).loadingTime).toEqual(FAKE_NAVIGATION_ENTRY.loadEventEnd)
+  })
+
+  it('should use loadEventEnd for initial view when load event is bigger than computed loading time', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+    expect(getRumEventCount()).toEqual(1)
+
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+
+    lifeCycle.notify(
+      LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED,
+      FAKE_NAVIGATION_ENTRY_WITH_LOADEVENT_AFTER_ACTIVITY_TIMING as PerformanceNavigationTiming
+    )
+
+    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    clock.tick(AFTER_PAGE_ACTIVITY_END_DELAY)
+
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(getRumEventCount()).toEqual(2)
+    expect(getViewEvent(1).loadingTime).toEqual(FAKE_NAVIGATION_ENTRY_WITH_LOADEVENT_AFTER_ACTIVITY_TIMING.loadEventEnd)
+  })
+
+  it('should use computed loading time for initial view when load event is smaller than computed loading time', () => {
+    const { lifeCycle, clock } = setupBuilder.build()
+    expect(getRumEventCount()).toEqual(1)
+
+    clock.tick(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+    lifeCycle.notify(
+      LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED,
+      FAKE_NAVIGATION_ENTRY_WITH_LOADEVENT_BEFORE_ACTIVITY_TIMING as PerformanceNavigationTiming
+    )
+    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    clock.tick(AFTER_PAGE_ACTIVITY_END_DELAY)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(getRumEventCount()).toEqual(2)
+    expect(getViewEvent(1).loadingTime).toEqual(BEFORE_PAGE_ACTIVITY_VALIDATION_DELAY)
+  })
+})
+
 describe('rum view measures', () => {
-  const FAKE_LONG_TASK = {
-    entryType: 'longtask',
-    startTime: 456,
-  }
-  const FAKE_USER_ACTION = {
-    context: {
-      bar: 123,
-    },
-    name: 'foo',
-    type: UserActionType.CUSTOM,
-  }
-  const FAKE_PAINT_ENTRY = {
-    entryType: 'paint',
-    name: 'first-contentful-paint',
-    startTime: 123,
-  }
-  const FAKE_NAVIGATION_ENTRY = {
-    domComplete: 456,
-    domContentLoadedEventEnd: 345,
-    domInteractive: 234,
-    entryType: 'navigation',
-    loadEventEnd: 567,
-  }
-  let lifeCycle: LifeCycle
+  let setupBuilder: TestSetupBuilder
+  let addRumEvent: jasmine.Spy
   let getRumEventCount: () => number
   let getViewEvent: (index: number) => View
+
   beforeEach(() => {
-    ;({ lifeCycle, getRumEventCount, getViewEvent } = spyOnViews())
+    ;({ addRumEvent, getViewEvent, getRumEventCount } = spyOnViews())
+
+    const fakeLocation: Partial<Location> = { pathname: '/foo' }
+    mockHistory(fakeLocation)
+    setupBuilder = setup()
+      .withViewCollection(fakeLocation)
+      .beforeBuild((lifeCycle) => lifeCycle.subscribe(LifeCycleEventType.VIEW_COLLECTED, addRumEvent))
+  })
+
+  afterEach(() => {
+    setupBuilder.cleanup()
   })
 
   it('should track error count', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures.errorCount).toEqual(0)
 
@@ -144,6 +325,7 @@ describe('rum view measures', () => {
   })
 
   it('should track long task count', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures.longTaskCount).toEqual(0)
 
@@ -156,6 +338,7 @@ describe('rum view measures', () => {
   })
 
   it('should track resource count', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures.resourceCount).toEqual(0)
 
@@ -168,6 +351,7 @@ describe('rum view measures', () => {
   })
 
   it('should track user action count', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures.userActionCount).toEqual(0)
 
@@ -180,6 +364,7 @@ describe('rum view measures', () => {
   })
 
   it('should reset event count when the view changes', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures.resourceCount).toEqual(0)
 
@@ -200,7 +385,7 @@ describe('rum view measures', () => {
   })
 
   it('should update measures when notified with a PERFORMANCE_ENTRY_COLLECTED event (throttled)', () => {
-    jasmine.clock().install()
+    const { lifeCycle, clock } = setupBuilder.withFakeClock().build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures).toEqual({
       errorCount: 0,
@@ -216,7 +401,7 @@ describe('rum view measures', () => {
 
     expect(getRumEventCount()).toEqual(1)
 
-    jasmine.clock().tick(THROTTLE_VIEW_UPDATE_PERIOD)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getRumEventCount()).toEqual(2)
     expect(getViewEvent(1).measures).toEqual({
@@ -229,12 +414,10 @@ describe('rum view measures', () => {
       resourceCount: 0,
       userActionCount: 0,
     })
-
-    jasmine.clock().uninstall()
   })
 
   it('should update measures when notified with a RESOURCE_ADDED_TO_BATCH event (throttled)', () => {
-    jasmine.clock().install()
+    const { lifeCycle, clock } = setupBuilder.withFakeClock().build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures).toEqual({
       errorCount: 0,
@@ -247,7 +430,7 @@ describe('rum view measures', () => {
 
     expect(getRumEventCount()).toEqual(1)
 
-    jasmine.clock().tick(THROTTLE_VIEW_UPDATE_PERIOD)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getRumEventCount()).toEqual(2)
     expect(getViewEvent(1).measures).toEqual({
@@ -256,11 +439,10 @@ describe('rum view measures', () => {
       resourceCount: 1,
       userActionCount: 0,
     })
-
-    jasmine.clock().uninstall()
   })
 
   it('should update measures when ending a view', () => {
+    const { lifeCycle } = setupBuilder.build()
     expect(getRumEventCount()).toEqual(1)
     expect(getViewEvent(0).measures).toEqual({
       errorCount: 0,
@@ -299,7 +481,7 @@ describe('rum view measures', () => {
   })
 
   it('should not update measures after ending a view', () => {
-    jasmine.clock().install()
+    const { lifeCycle, clock } = setupBuilder.withFakeClock().build()
     expect(getRumEventCount()).toEqual(1)
 
     lifeCycle.notify(LifeCycleEventType.RESOURCE_ADDED_TO_BATCH)
@@ -312,10 +494,8 @@ describe('rum view measures', () => {
     expect(getViewEvent(1).id).toEqual(getViewEvent(0).id)
     expect(getViewEvent(2).id).not.toEqual(getViewEvent(0).id)
 
-    jasmine.clock().tick(THROTTLE_VIEW_UPDATE_PERIOD)
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getRumEventCount()).toEqual(3)
-
-    jasmine.clock().uninstall()
   })
 })